merge new version

+++ ci/do_circle_ci.sh
@@ -4,7 +4,9 @@ set -e
 
 # bazel uses jgit internally and the default circle-ci .gitconfig says to
 # convert https://github.com to ssh://git@github.com, which jgit does not support.
-mv ~/.gitconfig ~/.gitconfig_save
+if [[ -e "~/.gitconfig" ]]; then
+  mv ~/.gitconfig ~/.gitconfig_save
+fi
 
 export ENVOY_SRCDIR="$(pwd)"
 
@@ -27,4 +29,4 @@ trap finish EXIT
 echo "disk space at beginning of build:"
 df -h
 
-ci/do_ci.sh "$1"
+ci/do_ci.sh $*
